+++ src/index.ts
@@ -4,16 +4,11 @@ import skosConceptScheme from './skosConceptScheme'
 import baseIri from './objects/baseIri'
 import TreeView from './components/TreeView'
 
-
 const taxonomyManager = definePlugin({
   name: 'taxonomyManager',
   schema: {
-    types: [
-      skosConcept, 
-      skosConceptScheme,
-      baseIri
-    ]
-  }
+    types: [skosConcept, skosConceptScheme, baseIri],
+  },
 })
 
 export {taxonomyManager, TreeView}