+++ src/types.ts
@@ -1,19 +1,19 @@
 /**
- * @todo Troubleshoot PrefLableValue interface declaration. `components: {input: }` in skosConcept.tsx doesn't like it. 
+ * @todo Troubleshoot PrefLableValue interface declaration. `components: {input: }` in skosConcept.tsx doesn't like it.
  */
 
 import {SanityDocument} from '@sanity/client'
 
 export interface TopConceptTerms {
-  prefLabel: string;
-  id: string;
-  childConcepts?: ChildConceptTerms[];
+  prefLabel: string
+  id: string
+  childConcepts?: ChildConceptTerms[]
 }
 
 export interface ChildConceptTerms {
-  prefLabel: string;
-  id: string;
-  childConcepts?: ChildConceptTerms[];
+  prefLabel: string
+  id: string
+  childConcepts?: ChildConceptTerms[]
 }
 
 export interface DocumentVersionsCollection {
@@ -24,6 +24,6 @@ export interface DocumentVersionsCollection {
 }
 
 export interface PrefLabelValue {
-  value: string;
+  value: string
   renderDefault: (props: PrefLabelValue) => React.ReactElement
 }